+++ hrefs/__init__.py
@@ -1,6 +1,6 @@
 """Hyperlinks for pydantic models"""
 
-__version__ = "0.10"
+__version__ = "0.11"
 
 from .href import Href, Referrable
 from .model import BaseReferrableModel, PrimaryKey